+++ FintrakBanking.Entities/Models/TBL_LC_USSANCE.cs
@@ -18,7 +18,7 @@
 
         public decimal USSANCEAMOUNT { get; set; }
 
-        public int? USSANCERATE { get; set; }
+        public decimal? USSANCERATE { get; set; }
 
         public virtual TBL_LC_ISSUANCE TBL_LC_ISSUANCE { get; set; }
